Linear System Solvers

solve(prob::LinearProblem,alg;kwargs)

Solves for $Au=b$ in the problem defined by prob using the algorithm alg. If no algorithm is given, a default algorithm will be chosen.

Recommended Methods

The default algorithm nothing is good for picking an algorithm that will work, but one may need to change this to receive more performance or precision. If more precision is necessary, QRFactorization() and SVDFactorization() are the best choices, with SVD being the slowest but most precise.

For efficiency, RFLUFactorization is the fastest for dense LU-factorizations. FastLUFactorization will be faster than LUFactorization which is the Base.LinearAlgebra (\ default) implementation of LU factorization. SimpleLUFactorization will be fast on very small matrices.

For sparse LU-factorizations, KLUFactorization if there is less structure to the sparsity pattern and UMFPACKFactorization if there is more structure. Pardiso.jl's methods are also known to be very efficient sparse linear solvers.

While these sparse factorizations are based on implementations in other languages, and therefore constrained to standard number types (Float64, Float32 and their complex counterparts), SparspakFactorization is able to handle general number types, e.g. defined by ForwardDiff.jl, MultiFloats.jl, or IntervalArithmetics.jl.

As sparse matrices get larger, iterative solvers tend to get more efficient than factorization methods if a lower tolerance of the solution is required.

Krylov.jl generally outperforms IterativeSolvers.jl and KrylovKit.jl, and is compatible with CPUs and GPUs, and thus is the generally preferred form for Krylov methods.

Finally, a user can pass a custom function for handling the linear solve using LinearSolveFunction() if existing solvers are not optimally suited for their application. The interface is detailed here.

Full List of Methods

RecursiveFactorization.jl

LinearSolve.RFLUFactorization — Type

RFLUFactorization()

A fast pure Julia LU-factorization implementation using RecursiveFactorization.jl. This is by far the fastest LU-factorization implementation, usually outperforming OpenBLAS and MKL for smaller matrices (<500x500), but currently optimized only for Base Array with Float32 or Float64. Additional optimization for complex matrices is in the works.

Base.LinearAlgebra

These overloads tend to work for many array types, such as CuArrays for GPU-accelerated solving, using the overloads provided by the respective packages. Given that this can be customized per-package, details given below describe a subset of important arrays (Matrix, SparseMatrixCSC, CuMatrix, etc.)

LinearSolve.LUFactorization — Type

LUFactorization(pivot=LinearAlgebra.RowMaximum())

Julia's built in lu. Equivalent to calling lu!(A)

On dense matrices, this uses the current BLAS implementation of the user's computer,

which by default is OpenBLAS but will use MKL if the user does using MKL in their system.

On sparse matrices, this will use UMFPACK from SuiteSparse. Note that this will not

cache the symbolic factorization.

On CuMatrix, it will use a CUDA-accelerated LU from CuSolver.

On BandedMatrix and BlockBandedMatrix, it will use a banded LU.

Positional Arguments

pivot: The choice of pivoting. Defaults to LinearAlgebra.RowMaximum(). The other choice is LinearAlgebra.NoPivot().

LinearSolve.GenericLUFactorization — Type

GenericLUFactorization(pivot=LinearAlgebra.RowMaximum())

Julia's built in generic LU factorization. Equivalent to calling LinearAlgebra.generic_lufact!. Supports arbitrary number types but does not achieve as good scaling as BLAS-based LU implementations. Has low overhead and is good for small matrices.

Positional Arguments

pivot: The choice of pivoting. Defaults to LinearAlgebra.RowMaximum(). The other choice is LinearAlgebra.NoPivot().

LinearSolve.QRFactorization — Type

QRFactorization(pivot=LinearAlgebra.NoPivot(),blocksize=16)

Julia's built in qr. Equivalent to calling qr!(A).

On dense matrices, this uses the current BLAS implementation of the user's computer

which by default is OpenBLAS but will use MKL if the user does using MKL in their system.

On sparse matrices, this will use SPQR from SuiteSparse

On CuMatrix, it will use a CUDA-accelerated QR from CuSolver.

On BandedMatrix and BlockBandedMatrix, it will use a banded QR.

LinearSolve.SVDFactorization — Type

SVDFactorization(full=false,alg=LinearAlgebra.DivideAndConquer())

Julia's built in svd. Equivalent to svd!(A).

On dense matrices, this uses the current BLAS implementation of the user's computer

which by default is OpenBLAS but will use MKL if the user does using MKL in their system.

LinearSolve.CholeskyFactorization — Type

CholeskyFactorization(; pivot = nothing, tol = 0.0, shift = 0.0, perm = nothing)

Julia's built in cholesky. Equivalent to calling cholesky!(A).

Keyword Arguments

pivot: defaluts to NoPivot, can also be RowMaximum.

tol: the tol argument in CHOLMOD. Only used for sparse matrices.

shift: the shift argument in CHOLMOD. Only used for sparse matrices.

perm: the perm argument in CHOLMOD. Only used for sparse matrices.

LinearSolve.BunchKaufmanFactorization — Type

BunchKaufmanFactorization(; rook = false)

Julia's built in bunchkaufman. Equivalent to calling bunchkaufman(A). Only for Symmetric matrices.

Keyword Arguments

rook: whether to perform rook pivoting. Defaults to false.

LinearSolve.CHOLMODFactorization — Type

CHOLMODFactorization(; shift = 0.0, perm = nothing)

A wrapper of CHOLMOD's polyalgorithm, mixing Cholesky factorization and ldlt. Tries cholesky for performance and retries ldlt if conditioning causes Cholesky to fail.

Only supports sparse matrices.

Keyword Arguments

shift: the shift argument in CHOLMOD.

perm: the perm argument in CHOLMOD

LinearSolve.NormalCholeskyFactorization — Type

NormalCholeskyFactorization(pivot = RowMaximum())

A fast factorization which uses a Cholesky factorization on A * A'. Can be much faster than LU factorization, but is not as numerically stable and thus should only be applied to well-conditioned matrices.

Positional Arguments

pivot: Defaults to RowMaximum(), but can be NoPivot()

LinearSolve.NormalBunchKaufmanFactorization — Type

NormalBunchKaufmanFactorization(rook = false)

A fast factorization which uses a BunchKaufman factorization on A * A'. Can be much faster than LU factorization, but is not as numerically stable and thus should only be applied to well-conditioned matrices.

Positional Arguments

rook: whether to perform rook pivoting. Defaults to false.

LinearSolve.jl

LinearSolve.jl contains some linear solvers built in for specailized cases.

LinearSolve.SimpleLUFactorization — Type

SimpleLUFactorization(pivot::Bool = true)

A simple LU-factorization implementation without BLAS. Fast for small matrices.

Positional Arguments

pivot::Bool: whether to perform pivoting. Defaults to true

LinearSolve.DiagonalFactorization — Type

DiagonalFactorization()

A special implementation only for solving Diagonal matrices fast.

FastLapackInterface.jl

FastLapackInterface.jl is a package that allows for a lower-level interface to the LAPACK calls to allow for preallocating workspaces to decrease the overhead of the wrappers. LinearSolve.jl provides a wrapper to these routines in a way where an initialized solver has a non-allocating LU factorization. In theory, this post-initialized solve should always be faster than the Base.LinearAlgebra version.

LinearSolve.FastLUFactorization — Type

FastLUFactorization()

The FastLapackInterface.jl version of the LU factorization. Notably, this version does not allow for choice of pivoting method.

LinearSolve.FastQRFactorization — Type

FastQRFactorization()

The FastLapackInterface.jl version of the QR factorization.

SuiteSparse.jl

LinearSolve.KLUFactorization — Type

KLUFactorization(;reuse_symbolic=true, check_pattern=true)

A fast sparse LU-factorization which specializes on sparsity patterns with “less structure”.

Note

By default, the SuiteSparse.jl are implemented for efficiency by caching the symbolic factorization. I.e., if set_A is used, it is expected that the new A has the same sparsity pattern as the previous A. If this algorithm is to be used in a context where that assumption does not hold, set reuse_symbolic=false.

LinearSolve.UMFPACKFactorization — Type

UMFPACKFactorization(;reuse_symbolic=true, check_pattern=true)

A fast sparse multithreaded LU-factorization which specializes on sparsity patterns with “more structure”.

Note

By default, the SuiteSparse.jl are implemented for efficiency by caching the symbolic factorization. I.e., if set_A is used, it is expected that the new A has the same sparsity pattern as the previous A. If this algorithm is to be used in a context where that assumption does not hold, set reuse_symbolic=false.

Sparspak.jl

LinearSolve.SparspakFactorization — Type

SparspakFactorization(reuse_symbolic = true)

This is the translation of the well-known sparse matrix software Sparspak (Waterloo Sparse Matrix Package), solving large sparse systems of linear algebraic equations. Sparspak is composed of the subroutines from the book "Computer Solution of Large Sparse Positive Definite Systems" by Alan George and Joseph Liu. Originally written in Fortran 77, later rewritten in Fortran 90. Here is the software translated into Julia.

The Julia rewrite is released under the MIT license with an express permission from the authors of the Fortran package. The package uses multiple dispatch to route around standard BLAS routines in the case e.g. of arbitrary-precision floating point numbers or ForwardDiff.Dual. This e.g. allows for Automatic Differentiation (AD) of a sparse-matrix solve.

Krylov.jl

LinearSolve.KrylovJL_CG — Function

KrylovJL_CG(args...;  kwargs...)

A generic CG implementation for Hermitian and positive definite linear systems

LinearSolve.KrylovJL_MINRES — Function

KrylovJL_MINRES(args...;  kwargs...)

A generic MINRES implementation for Hermitian linear systems

LinearSolve.KrylovJL_GMRES — Function

KrylovJL_GMRES(args...;  gmres_restart = 0, window = 0, kwargs...)

A generic GMRES implementation for square non-Hermitian linear systems

LinearSolve.KrylovJL_BICGSTAB — Function

KrylovJL_BICGSTAB(args...;  kwargs...)

A generic BICGSTAB implementation for square non-Hermitian linear systems

LinearSolve.KrylovJL_LSMR — Function

KrylovJL_LSMR(args...;  kwargs...)

A generic LSMR implementation for least-squares problems

LinearSolve.KrylovJL_CRAIGMR — Function

KrylovJL_CRAIGMR(args...;  kwargs...)

A generic CRAIGMR implementation for least-norm problems

LinearSolve.KrylovJL — Type

KrylovJL(args...; KrylovAlg = Krylov.gmres!,
         Pl = nothing, Pr = nothing,
         gmres_restart = 0, window = 0,
         kwargs...)

A generic wrapper over the Krylov.jl krylov-subspace iterative solvers.

Pardiso.jl

Note

Using this solver requires adding the package Pardiso.jl, i.e. using Pardiso

LinearSolve.MKLPardisoFactorize — Function

```julia MKLPardisoFactorize(; nprocs::Union{Int, Nothing} = nothing, matrix_type = nothing, iparm::Union{Vector{Tuple{Int, Int}}, Nothing} = nothing, dparm::Union{Vector{Tuple{Int, Int}}, Nothing} = nothing)

A sparse factorization method using MKL Pardiso.

Note

Using this solver requires adding the package Pardiso.jl, i.e. using Pardiso

Keyword Arguments

For the definition of the keyword arguments, see the Pardiso.jl documentation. All values default to nothing and the solver internally determines the values given the input types, and these keyword arguments are only for overriding the default handling process. This should not be required by most users.

LinearSolve.MKLPardisoIterate — Function

MKLPardisoIterate(; nprocs::Union{Int, Nothing} = nothing,
                    matrix_type = nothing,
                    iparm::Union{Vector{Tuple{Int, Int}}, Nothing} = nothing,
                    dparm::Union{Vector{Tuple{Int, Int}}, Nothing} = nothing)

A mixed factorization+iterative method using MKL Pardiso.

Note

Using this solver requires adding the package Pardiso.jl, i.e. using Pardiso

Keyword Arguments

For the definition of the keyword arguments, see the Pardiso.jl documentation. All values default to nothing and the solver internally determines the values given the input types, and these keyword arguments are only for overriding the default handling process. This should not be required by most users.

Missing docstring.

Missing docstring for LinearSolve.PardisoJL. Check Documenter's build log for details.

CUDA.jl

Note that CuArrays are supported by GenericFactorization in the “normal” way. The following are non-standard GPU factorization routines.

Note

Using this solver requires adding the package CUDA.jl, i.e. using CUDA

LinearSolve.CudaOffloadFactorization — Type

CudaOffloadFactorization()

An offloading technique used to GPU-accelerate CPU-based computations. Requires a sufficiently large A to overcome the data transfer costs.

Note

Using this solver requires adding the package CUDA.jl, i.e. using CUDA

IterativeSolvers.jl

Note

Using these solvers requires adding the package IterativeSolvers.jl, i.e. using IterativeSolvers

LinearSolve.IterativeSolversJL_CG — Function

IterativeSolversJL_CG(args...; Pl = nothing, Pr = nothing, kwargs...)

A wrapper over the IterativeSolvers.jl CG.

Note

Using this solver requires adding the package IterativeSolvers.jl, i.e. using IterativeSolvers

LinearSolve.IterativeSolversJL_GMRES — Function

IterativeSolversJL_GMRES(args...; Pl = nothing, Pr = nothing, gmres_restart=0, kwargs...)

A wrapper over the IterativeSolvers.jl GMRES.

Note

Using this solver requires adding the package IterativeSolvers.jl, i.e. using IterativeSolvers

LinearSolve.IterativeSolversJL_BICGSTAB — Function

IterativeSolversJL_BICGSTAB(args...; Pl = nothing, Pr = nothing, kwargs...)

A wrapper over the IterativeSolvers.jl BICGSTAB.

Note

Using this solver requires adding the package IterativeSolvers.jl, i.e. using IterativeSolvers

LinearSolve.IterativeSolversJL_MINRES — Function

IterativeSolversJL_MINRES(args...; Pl = nothing, Pr = nothing, kwargs...)

A wrapper over the IterativeSolvers.jl MINRES.

Note

Using this solver requires adding the package IterativeSolvers.jl, i.e. using IterativeSolvers

LinearSolve.IterativeSolversJL — Type

IterativeSolversJL(args...;
                   generate_iterator = IterativeSolvers.gmres_iterable!,
                   Pl = nothing, Pr = nothing,
                   gmres_restart = 0, kwargs...)

A generic wrapper over the IterativeSolvers.jl solvers.

Note

Using this solver requires adding the package IterativeSolvers.jl, i.e. using IterativeSolvers

KrylovKit.jl

Note

Using these solvers requires adding the package KrylovKit.jl, i.e. using KrylovKit

LinearSolve.KrylovKitJL_CG — Function

KrylovKitJL_CG(args...; Pl = nothing, Pr = nothing, kwargs...)

A generic CG implementation for Hermitian and positive definite linear systems

Note

Using this solver requires adding the package KrylovKit.jl, i.e. using KrylovKit

LinearSolve.KrylovKitJL_GMRES — Function

KrylovKitJL_GMRES(args...; Pl = nothing, Pr = nothing, gmres_restart = 0, kwargs...)

A generic GMRES implementation.

Note

Using this solver requires adding the package KrylovKit.jl, i.e. using KrylovKit

LinearSolve.KrylovKitJL — Type

KrylovKitJL(args...; KrylovAlg = Krylov.gmres!, kwargs...)

A generic iterative solver implementation allowing the choice of KrylovKit.jl solvers.

Note

Using this solver requires adding the package KrylovKit.jl, i.e. using KrylovKit

HYPRE.jl

Note

Using HYPRE solvers requires Julia version 1.9 or higher, and that the package HYPRE.jl is installed.

LinearSolve.HYPREAlgorithm — Type

HYPREAlgorithm(solver; Pl = nothing)

HYPRE.jl is an interface to hypre and provide iterative solvers and preconditioners for sparse linear systems. It is mainly developed for large multi-process distributed problems (using MPI), but can also be used for single-process problems with Julias standard sparse matrices.

If you need more fine-grained control over the solver/preconditioner options you can alternatively pass an already created solver to HYPREAlgorithm (and to the Pl keyword argument). See HYPRE.jl docs for how to set up solvers with specific options.

Note

Using HYPRE solvers requires Julia version 1.9 or higher, and that the package HYPRE.jl is installed.

Positional Arguments

The single positional argument solver has the following choices:

HYPRE.BiCGSTAB

HYPRE.BoomerAMG

HYPRE.FlexGMRES

HYPRE.GMRES

HYPRE.Hybrid

HYPRE.ILU

HYPRE.ParaSails (as preconditioner only)

HYPRE.PCG

Keyword Arguments

Pl: A choice of left preconditioner.

Example

For example, to use HYPRE.PCG as the solver, with HYPRE.BoomerAMG as the preconditioner, the algorithm should be defined as follows:

A, b = setup_system(...)
prob = LinearProblem(A, b)
alg = HYPREAlgorithm(HYPRE.PCG)
prec = HYPRE.BoomerAMG
sol = solve(prob, alg; Pl = prec)